API Documentation

Octofitter.Priors — Type

Priors(key=dist, ...)

A group of zero or more priors passed by keyword arguments. The priors must be univariate distributions from the Distributions.jl package.

Octofitter.Derived — Type

Derived(key=func, ...)

A group of zero or more functions that resolve to a parameter for the model. Derived parameters must be functions accepting one argument if applied to a System, or two arguments if applied to a Planet. Must always return the same value for the same input (be a pure function) and support autodifferentiation.

Missing docstring.

Missing docstring for AstrometryLikelihood. Check Documenter's build log for details.

Missing docstring.

Missing docstring for PhotometryLikelihood. Check Documenter's build log for details.

Missing docstring.

Missing docstring for ProperMotionAnomLikelihood. Check Documenter's build log for details.

Octofitter.HGCALikelihood — Type

HGCALikelihood(;gaia_id=1234)

Load proper motion anomaly data from the HIPPARCOS-GAIA Catalog of Accelerations (Brandt et al) for a star with catalog id gaia_id. The resulting velocities are in mas/yr and have the long term trend between HIPPARCOS and GAIA already subtracted out. e.g. we would expect 0 pma if there is no companion.

Missing docstring.

Missing docstring for ImageLikelihood. Check Documenter's build log for details.

Octofitter.Sine — Type

Sine()

A custom univariate distribution. The pdf is a sine function defined between 0 and π. This is a common prior distribution used when fitting orbits to astrometry.

The full Distributions.jl interface is not yet defined for this distribution, but the following methods work: pdf, logpdf, minimum, maximum, insupport, mean, var, cdf, quantile

Octofitter.Planet — Type

Planet([derived,] priors, [astrometry,], name=:symbol)

A planet (or substellar companion) part of a model. Must be constructed with a block of priors, and optionally additional derived parameters and/or sastrometry. name must be a symbol, e.g. :b.

Octofitter.System — Type

System([derived,] priors, [images,] [propermotionanom,] planets..., name=:symbol)

Construct a model of a system. Must be constructed with a block of priors, and optionally additional derived parameters. You may provide ProperMotionAnomLikelihood() and/or Images() of the system. Finally, planet models are listed last. name must be a symbol e.g. :HD56441.

PlanetOrbits.mjd — Function

mjd("2020-01-01")

Get the modfied julian day of a date, or in general a UTC timestamp.

mjd()

Get the current modified julian day of right now.

Octofitter.gaia_plx — Function

gaia_plx(gaia_id=12123)

Get a distribution (truncated Normal) of parallax distance in mas of a source with GAIA catalog id gaia_id.

Missing docstring.

Missing docstring for Octofitter.hmc. Check Documenter's build log for details.

PlanetOrbits.Visual — Type

Visual{OrbitType}(..., plx=...)

This wraps another orbit to add the parallax distance field plx, thus allowing projected quantities to be calculated. It forwards everything else to the parent orbit.

For example, the KepOrbit type supports calculating x and y positions in AU. A Visual{KepOrbit} additionally supports calculating projected right ascension and declination offsets.

Note

The ThieleInnesOrbit type does not need to be wrapped in Visual as it the Thiele-Innes constants are already expressed in milliarcseconds and thus it always requires a plx value.

Octofitter.sonora_photometry_interpolator — Function

sonora_photometry_interpolator(:Keck_L′, [metalicity="+0.0"])

Given a supported photometric band and [M/H] metalicity (default=solar), return a function of temperature (K) and mass (M_jup) that gives the absolute magnitude of the planet in that bandpass.

Supported bands: :MKOY, :MKOZ, :MKOJ, :MKOH, :MKOK, :MKOL′, :MKOM′, :TwoMASSJ, :TwoMASSH, :TwoMASSKs, :KeckKs, :KeckL′, :KeckMs, :SDSSg′, :SDSSr′, :SDSSi′, :SDSSz′, :IRAC36, :IRAC45, :IRAC57, :IRAC79, :WISEW1, :WISEW2, :WISEW3, :WISE_W4

Supported metalicities: "+0.0", "-0.5", "+0.5"

Missing docstring.

Missing docstring for sonora_cooling_interpolator. Check Documenter's build log for details.

Missing docstring.

Missing docstring for plotchains. Check Documenter's build log for details.

Octofitter.projectpositions — Function

projectpositions(chains.planets[1], mjd("2020-02-02"))

Given the posterior for a particular planet in the model and a modified julian date(s), return ra and dec offsets in mas for each sampling in the posterior.

Missing docstring.

Missing docstring for generate. Check Documenter's build log for details.